fix(tasks): code cleanup
